+++ CHANGES.md
@@ -33,6 +33,7 @@
     * Passwords are no longer kept in memory by the plugin, but instead
       use the IDE's password storage mechanism.  Before, only passwords
       that the user entered through the dialog were stored by the IDE.
+    * File annotations should be faster now.
 * Active Connection view.
     * Rather than using a small status-bar widget, connections are now
       viewed through a VCS tab called "Active Connections".